Added silent function to allow execution of commands without running cmd.exe

+++ examples/atexec.py
@@ -32,7 +32,7 @@ from impacket.dcerpc.v5.rpcrt import RPC_C_AUTHN_GSS_NEGOTIATE
 
 class TSCH_EXEC:
     def __init__(self, username='', password='', domain='', hashes=None, aesKey=None, doKerberos=False, kdcHost=None,
-                 command=None):
+                 command=None, silentCommand=False):
         self.__username = username
         self.__password = password
         self.__domain = domain
@@ -42,6 +42,7 @@ class TSCH_EXEC:
         self.__doKerberos = doKerberos
         self.__kdcHost = kdcHost
         self.__command = command
+        self.__silentCommand = silentCommand
         if hashes is not None:
             self.__lmhash, self.__nthash = hashes.split(':')
 
@@ -116,12 +117,13 @@ class TSCH_EXEC:
   </Settings>
   <Actions Context="LocalSystem">
     <Exec>
-      <Command>cmd.exe</Command>
-      <Arguments>/C %s &gt; %%windir%%\\Temp\\%s 2&gt;&amp;1</Arguments>
+      <Command>%s</Command>
+      <Arguments>%s</Arguments>
     </Exec>
   </Actions>
 </Task>
-        """ % (self.__command, tmpFileName)
+        """ % (("cmd.exe" if self.__silentCommand is False else self.__command.split()[0]), 
+            (("/C %s &gt; %%windir%%\\Temp\\%s 2&gt;&amp;1" % (self.__command, tmpFileName)) if self.__silentCommand is False else " ".join(self.__command.split()[1:])))
         taskCreated = False
         try:
             logging.info('Creating task \\%s' % tmpName)
@@ -150,27 +152,28 @@ class TSCH_EXEC:
             if taskCreated is True:
                 tsch.hSchRpcDelete(dce, '\\%s' % tmpName)
 
-        smbConnection = rpctransport.get_smb_connection()
-        waitOnce = True
-        while True:
-            try:
-                logging.info('Attempting to read ADMIN$\\Temp\\%s' % tmpFileName)
-                smbConnection.getFile('ADMIN$', 'Temp\\%s' % tmpFileName, output_callback)
-                break
-            except Exception as e:
-                if str(e).find('SHARING') > 0:
-                    time.sleep(3)
-                elif str(e).find('STATUS_OBJECT_NAME_NOT_FOUND') >= 0:
-                    if waitOnce is True:
-                        # We're giving it the chance to flush the file before giving up
+        if not self.__silentCommand:
+            smbConnection = rpctransport.get_smb_connection()
+            waitOnce = True
+            while True:
+                try:
+                    logging.info('Attempting to read ADMIN$\\Temp\\%s' % tmpFileName)
+                    smbConnection.getFile('ADMIN$', 'Temp\\%s' % tmpFileName, output_callback)
+                    break
+                except Exception as e:
+                    if str(e).find('SHARING') > 0:
                         time.sleep(3)
-                        waitOnce = False
+                    elif str(e).find('STATUS_OBJECT_NAME_NOT_FOUND') >= 0:
+                        if waitOnce is True:
+                            # We're giving it the chance to flush the file before giving up
+                            time.sleep(3)
+                            waitOnce = False
+                        else:
+                            raise
                     else:
                         raise
-                else:
-                    raise
-        logging.debug('Deleting file ADMIN$\\Temp\\%s' % tmpFileName)
-        smbConnection.deleteFile('ADMIN$', 'Temp\\%s' % tmpFileName)
+            logging.debug('Deleting file ADMIN$\\Temp\\%s' % tmpFileName)
+            smbConnection.deleteFile('ADMIN$', 'Temp\\%s' % tmpFileName)
  
         dce.disconnect()
 
@@ -187,6 +190,8 @@ if __name__ == '__main__':
 
     parser.add_argument('target', action='store', help='[[domain/]username[:password]@]<targetName or address>')
     parser.add_argument('command', action='store', nargs='*', default = ' ', help='command to execute at the target ')
+    parser.add_argument('-silentcommand', action='store_true', default = False, help='does not execute cmd.exe to run given command '
+                                                                                '(cannot run dir/cd/etc.)')
     parser.add_argument('-debug', action='store_true', help='Turn DEBUG output ON')
 
     group = parser.add_argument_group('authentication')
@@ -238,5 +243,5 @@ if __name__ == '__main__':
         options.k = True
 
     atsvc_exec = TSCH_EXEC(username, password, domain, options.hashes, options.aesKey, options.k, options.dc_ip,
-                           ' '.join(options.command))
+                           ' '.join(options.command), options.silentcommand)
     atsvc_exec.play(address)
